feat(git): add branc_bump
This function adds a -2 (or -3, -4...) to the current branch name, both locally and upstream. Use case:
When using gh to create a branch linked to an issue, it will always create the same branch name. But if you have already merged a PR with a branch of that name, things will get messy.